// src/modules/voucher/voucher.enum.ts
export enum VoucherType {
  PERCENT = "PERCENT",
  FIXED = "FIXED",
}

export enum VoucherFieldName {
  CODE = "code",
  FRANCHISE_ID = "franchise_id",
  PRODUCT_FRANCHISE_ID = "product_franchise_id",
  NAME = "name",
  DESCRIPTION = "description",
  TYPE = "type",
  VALUE = "value",
  QUOTA_TOTAL = "quota_total",
  QUOTA_USED = "quota_used",
  START_DATE = "start_date",
  END_DATE = "end_date",
  CREATED_BY = "created_by",
  IS_ACTIVE = "is_active",
  IS_DELETED = "is_deleted",
}